+++ collective/teamwork/user/groups.py
@@ -1,5 +1,7 @@
 from collections import OrderedDict
 import itertools
+import threading
+import weakref
 
 from zope.component import adapts
 from zope.component.hooks import getSite
@@ -15,6 +17,55 @@ _str = lambda v: v.encode('utf-8') if isinstance(v, unicode) else str(v)
 _u = lambda v: v.decode('utf-8') if isinstance(v, str) else unicode(v)
 
 
+class GroupInvalidation(threading.local):
+    """
+    Thread-local state synchronizer for GroupInfo cache invalidation.
+    We use this because shared state across transaction boundaries is more
+    complicated and risky (if transaction is aborted), and we use
+    thread-local storage as all other state is expected to comply with MVCC
+    via ZODB (as each thread and instance has own state/cache).
+    ** The goal is to have all state be consistent within a transaction. **
+
+    Impact on garbage collection:
+
+        - We use weakref to keep cached target references from preventing
+          GC on GroupInfo object.
+
+        - We use hash(target) to compute storage key, preventing duplicate
+          subscription.
+        
+        - GroupInfo.__del__() should unsubscribe from GroupInvalidation.
+    """
+
+    def __init__(self):
+        super(GroupInvalidation, self).__init__()
+        self.info = {}
+
+    def subscribe(self, name, target):
+        """Subscribe target object to invalidation for group name"""
+        if name not in self.info:
+            self.info[name] = {}
+        self.info[name][hash(target)] = weakref.ref(target)
+
+    def unsubscribe(self, name, target):
+        """Unsubscribe a target object to invalidation for group name"""
+        if name in self.info and hash(target) in self.info[name]:
+            del self.info[name][hash(target)]
+
+    def invalidate(self, name, context=None):
+        if name in self.info:
+            for ref in self.info[name].values():
+                signified = ref()
+                if signified is context:
+                    continue
+                target = ref()
+                if target is not None:
+                    target._usernames = None
+
+
+_group_invalidation = GroupInvalidation()
+
+
 class GroupInfo(object):
 
     implements(IGroup)
@@ -28,6 +79,13 @@ class GroupInfo(object):
         self._init_info()
         self._members = self._members_adapter(members)
         self.refresh()
+        _group_invalidation.subscribe(name, self)
+
+    def __del__(self):
+        _group_invalidation.unsubscribe(self._name, self)
+
+    def applyTransform(self, username):
+        return self._members.applyTransform(username)
 
     def _init_info(self):
         self._info = None
@@ -87,6 +145,7 @@ class GroupInfo(object):
 
     def refresh(self):
         self._usernames = None
+        _group_invalidation.invalidate(self._name, context=self)
 
     def keys(self):
         """User login name keys"""
@@ -124,6 +183,7 @@ class GroupInfo(object):
         return itertools.imap(_itemtuple, self.keys())
 
     def __contains__(self, username):
+        username = self.applyTransform(username)
         return username in self.keys()
 
     def __len__(self):
@@ -135,6 +195,7 @@ class GroupInfo(object):
         raise KeyError(username)
 
     def get(self, username, default=None):
+        username = self.applyTransform(username)
         if username in self.keys():
             return self._members.get(username)
         return default
@@ -147,6 +208,7 @@ class GroupInfo(object):
 
     def assign(self, username):
         """Add/assign a username to group"""
+        username = self.applyTransform(username)
         userid = self._members.userid_for(username)
         if userid is None:
             # possibly new user name, invalidate and try again
@@ -159,6 +221,7 @@ class GroupInfo(object):
 
     def unassign(self, username):
         """Unassign a username from a group"""
+        username = self.applyTransform(username)
         if username not in self.keys():
             raise ValueError('username provided is not in group')
         userid = self._members.userid_for(username)